+++ packages/plugin-stark-module/src/entryHelper.ts
@@ -1,0 +1,11 @@
+import { Json } from './types';
+
+export const getModules = (modules?: Json<string>[]) => {
+  return modules || {
+    'index': './src/index'
+  };
+};
+
+export const hasModules = (modules?: Json<string>[]) => {
+  return !!modules;
+};