+++ packages/build-mpa-config/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.2.1
+
+- chore: bump version
+
+## 3.2.0
+
+- [feat] support render custom tab bar in MPA
+- [fix] the page doesn't render when the app.ts entry exports a component in mpa
+
 ## 3.1.1
 
 - [chore] `hydrate` should be set when staticExport