+++ packages/build-user-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.4
+
+- [fix] remove deprecated api of esbuild-loader
+
+## 0.3.3
+
+- [feat] support resolve absolute cli options module path
+
 ## 0.3.2
 
 - [fix] custom config of terserOptions