+++ packages/plugin-app-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.4.3
+
+- [chore] ignore `unmount` function param never read error
+
+## 1.4.2
+
+- [feat] support resetPageStore param
+
 ## 1.4.1
 
 - [chore] set default value to false of generateRuntime