+++ packages/plugin-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.9.4
+
+- [fix] react-router-dom ts types not found
+
+## 1.9.3
+
+- [chore] ts types
+
 ## 1.9.2
 
 - [fix] join route path